+++ packages/kbn-test/src/functional_test_runner/lib/config/config.ts
@@ -58,7 +58,7 @@ export class Config {
     this[$values] = value;
   }
 
-  public has(key: string) {
+  public has(key: string | string[]) {
     function recursiveHasCheck(
       remainingPath: string[],
       values: Record<string, any>,
@@ -109,7 +109,7 @@ export class Config {
     return recursiveHasCheck(path, this[$values], schema);
   }
 
-  public get(key: string, defaultValue?: any) {
+  public get(key: string | string[], defaultValue?: any) {
     if (!this.has(key)) {
       throw new Error(`Unknown config key "${key}"`);
     }
